+++ parker/src/main/java/com/si9nal/parker/bookmark/domain/ViolationBookmark.java
@@ -1,7 +1,7 @@
 package com.si9nal.parker.bookmark.domain;
 
 import com.si9nal.parker.global.common.BaseEntity;
-import com.si9nal.parker.parkingvioation.domain.ParkingViolation;
+import com.si9nal.parker.parkingviolation.domain.ParkingViolation;
 import com.si9nal.parker.user.domain.User;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
