+++ parker/src/main/java/com/si9nal/parker/parkingviolation/repository/ParkingViolationRepository.java
@@ -1,6 +1,6 @@
-package com.si9nal.parker.parkingvioation.repository;
+package com.si9nal.parker.parkingviolation.repository;
 
-import com.si9nal.parker.parkingvioation.domain.ParkingViolation;
+import com.si9nal.parker.parkingviolation.domain.ParkingViolation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
